+++ CHANGELOG.md
@@ -266,3 +266,12 @@
 ## Release 0.0.1.9.3.1
 ----------------------
 - fixed open mode in save method in core.py (w->a)
+
+## Release 0.0.1.9.4.0
+----------------------
+### 01.03.2023
+- added KLD as compute_KLD method in util.py
+- added kwarg bins in compute_wasserstein_distance in util.py
+- added get_file_name method to util.py
+- added read_yaml mthod to core.py
+- fixed error in save_study in util.py (scalar lossed can now be saved)